+++ config/environments/development.rb
@@ -40,6 +40,8 @@ Rails.application.configure do
   # Set localhost to be used by links generated in mailer templates.
   config.action_mailer.default_url_options = { host: "localhost", port: 3000 }
 
+  MailerSettings.apply!(config) if MailerSettings.smtp_configured?
+
   # Print deprecation notices to the Rails logger.
   config.active_support.deprecation = :log